Cache put to update endpoint and make long type to received parameters

+++ src/main/java/com/germant/springPractice/repos/PersonRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 public interface PersonRepository extends CrudRepository<Person, Long> {
     List<Person> findAll();
 
-    Optional<Person> findById(Integer aLong);
+    Optional<Person> findById(Long aLong);
 
     Person save(Person person);
